refactor to avoid redundant code in view
