+++ expo/gbe/email/views/__init__.py
@@ -1,4 +1,5 @@
 from edit_template_view import EditTemplateView
 from list_template_view import ListTemplateView
+from mail_view import MailView
 from mail_to_bidders_view import MailToBiddersView
 from mail_to_person_view import MailToPersonView
