+++ expo/gbe/email/views/mail_to_bidders_view.py
@@ -1,16 +1,8 @@
-from django.views.generic import View
 from django.views.decorators.cache import never_cache
-from django.forms import HiddenInput
 from django.utils.decorators import method_decorator
 from django.contrib.auth.decorators import login_required
-from django.shortcuts import (
-    get_object_or_404,
-    render,
-)
-from django.http import (
-    Http404,
-    HttpResponseRedirect,
-)
+from django.shortcuts import render
+from django.http import HttpResponseRedirect
 from django.core.urlresolvers import reverse
 from django.contrib import messages
 from gbe.models import (
@@ -22,21 +14,19 @@ from gbe.models import (
     Volunteer,
 )
 from gbe.email.forms import (
-    AdHocEmailForm,
     SecretBidderInfoForm,
     SelectBidderForm,
 )
+from gbe.email.views import MailView
 from gbetext import (
-    send_email_success_msg,
     to_list_empty_msg,
     unknown_request,
 )
 from gbe.functions import validate_perms
 from django.db.models import Q
-from post_office import mail
 
 
-class MailToBiddersView(View):
+class MailToBiddersView(MailView):
     reviewer_permissions = ['Act Coordinator',
                             'Class Coordinator',
                             'Costume Coordinator',
@@ -100,10 +90,7 @@ class MailToBiddersView(View):
                 request,
                 'gbe/email/mail_to_bidders.tmpl',
                 {"selection_form": self.select_form})
-        email_form = AdHocEmailForm(initial={
-            'sender': self.user.user_object.email})
-        if not request.user.is_superuser:
-            email_form.fields['sender'].widget = HiddenInput()
+        email_form = self.setup_email_form(request)
         recipient_info = SecretBidderInfoForm(initial={
             'conference': self.select_form.cleaned_data['conference'],
             'bid_type': self.select_form.cleaned_data['bid_type'],
@@ -117,53 +104,6 @@ class MailToBiddersView(View):
             {"selection_form": self.select_form,
              "email_forms": [email_form, recipient_info],
              "to_list": to_list, })
-
-    def send_mail(self, request, to_list):
-        mail_form = AdHocEmailForm(request.POST)
-        if not request.user.is_superuser:
-            mail_form.fields['sender'].widget = HiddenInput()
-        recipient_info = SecretBidderInfoForm(request.POST,
-                                              prefix="email-select")
-        recipient_info.fields['bid_type'].choices = self.bid_type_choices
-        if mail_form.is_valid():
-            email_batch = []
-            recipient_string = ""
-            if request.user.is_superuser:
-                sender = mail_form.cleaned_data['sender']
-            else:
-                sender = request.user.email
-
-            for email, name in to_list.iteritems():
-                email_batch += [{
-                    'sender': sender,
-                    'recipients': [email],
-                    'subject': mail_form.cleaned_data['subject'],
-                    'html_message': mail_form.cleaned_data['html_message'], }]
-                recipient_string = "%s (%s), %s" % (
-                    name,
-                    email,
-                    recipient_string)
-
-            mail.send_many(email_batch)
-            user_message = UserMessage.objects.get_or_create(
-                view=self.__class__.__name__,
-                code="SEND_SUCCESS",
-                defaults={
-                    'summary': "Email Sent to Bidders",
-                    'description': send_email_success_msg})
-            messages.success(
-                request,
-                user_message[0].description + recipient_string)
-            return HttpResponseRedirect(
-                reverse('mail_to_bidders', urlconf='gbe.email.urls'))
-
-        else:
-            return render(
-                request,
-                'gbe/email/mail_to_bidders.tmpl',
-                {"selection_form": self.select_form,
-                 "email_forms": [mail_form, recipient_info],
-                 "to_list": to_list, })
 
     @never_cache
     @method_decorator(login_required)
@@ -183,7 +123,22 @@ class MailToBiddersView(View):
             return self.filter_bids(request)
         elif 'send' in request.POST.keys() and self.select_form.is_valid():
             to_list = self.get_to_list()
-            return self.send_mail(request, to_list)
+            mail_form = self.send_mail(request, to_list)
+            recipient_info = SecretBidderInfoForm(request.POST,
+                                              prefix="email-select")
+            recipient_info.fields['bid_type'].choices = self.bid_type_choices
+            if mail_form.is_valid():
+                return HttpResponseRedirect(
+                    reverse('mail_to_bidders', urlconf='gbe.email.urls'))
+
+            else:
+                return render(
+                    request,
+                    'gbe/email/mail_to_bidders.tmpl',
+                    {"selection_form": self.select_form,
+                     "email_forms": [mail_form, recipient_info],
+                     "to_list": to_list, })
+
         user_message = UserMessage.objects.get_or_create(
             view=self.__class__.__name__,
             code="UNKNOWN_ACTION",
@@ -198,7 +153,3 @@ class MailToBiddersView(View):
             'gbe/email/mail_to_bidders.tmpl',
             {"selection_form": self.select_form, }
             )
-
-    @method_decorator(login_required)
-    def dispatch(self, *args, **kwargs):
-        return super(MailToBiddersView, self).dispatch(*args, **kwargs)
